wireguard: timers: cast enum limits members to int in prints

+++ src/timers.c
@@ -48,8 +48,8 @@ static void wg_expired_retransmit_handshake(struct timer_list *timer)
 			  peer->endpoint.addr.sa_family == AF_INET &&
 			 !ipv4_is_zeronet(peer->endpoint.addr4.sin_addr.s_addr)) ||
 				peer->device->debug) {
-			net_info_peer_ratelimited("%s: handshake for peer \"%s\" (%llu) (%pISpfsc) did not complete after %lu attempts, giving up\n",
-				 peer, peer->internal_id, &peer->endpoint.addr, MAX_TIMER_HANDSHAKES + 2);
+			net_info_peer_ratelimited("%s: handshake for peer \"%s\" (%llu) (%pISpfsc) did not complete after %d attempts, giving up\n",
+				 peer, peer->internal_id, &peer->endpoint.addr, (int)MAX_TIMER_HANDSHAKES + 2);
 		}
 
 		del_timer(&peer->timer_send_keepalive);
@@ -70,8 +70,8 @@ static void wg_expired_retransmit_handshake(struct timer_list *timer)
 			  peer->endpoint.addr.sa_family == AF_INET &&
 			 !ipv4_is_zeronet(peer->endpoint.addr4.sin_addr.s_addr)) ||
 				peer->device->debug) {
-			net_info_peer_ratelimited("%s: handshake for peer \"%s\" (%llu) (%pISpfsc) did not complete after %lu seconds, retrying (try %d)\n",
-				 peer, peer->internal_id, &peer->endpoint.addr, REKEY_TIMEOUT,
+			net_info_peer_ratelimited("%s: handshake for peer \"%s\" (%llu) (%pISpfsc) did not complete after %d seconds, retrying (try %d)\n",
+				 peer, peer->internal_id, &peer->endpoint.addr, (int)REKEY_TIMEOUT,
 				 peer->timer_handshake_attempts + 1);
 		}
 
@@ -101,9 +101,9 @@ static void wg_expired_new_handshake(struct timer_list *timer)
 	struct wg_peer *peer = from_timer(peer, timer, timer_new_handshake);
 
 	if (peer->device->debug) {
-		net_info_peer_ratelimited("%s: retrying handshake with peer \"%s\" (%llu) (%pISpfsc) because we stopped hearing back after %lu seconds\n",
+		net_info_peer_ratelimited("%s: retrying handshake with peer \"%s\" (%llu) (%pISpfsc) because we stopped hearing back after %d seconds\n",
 			 peer, peer->internal_id,
-			 &peer->endpoint.addr, KEEPALIVE_TIMEOUT + REKEY_TIMEOUT);
+			 &peer->endpoint.addr, (int)(KEEPALIVE_TIMEOUT + REKEY_TIMEOUT));
 	}
 
 	/* We clear the endpoint address src address, in case this is the cause
@@ -136,9 +136,9 @@ static void wg_queued_expired_zero_key_material(struct work_struct *work)
 					    clear_peer_work);
 
 	if (peer->device->debug) {
-		net_info_peer_ratelimited("%s: zeroing out all keys for peer \"%s\" (%llu) (%pISpfsc), since we haven't received a new one in %lu seconds\n",
+		net_info_peer_ratelimited("%s: zeroing out all keys for peer \"%s\" (%llu) (%pISpfsc), since we haven't received a new one in %d seconds\n",
 			 peer, peer->internal_id,
-			&peer->endpoint.addr, REJECT_AFTER_TIME * 3);
+			&peer->endpoint.addr, (int)REJECT_AFTER_TIME * 3);
 	}
 
 	wg_noise_handshake_clear(&peer->handshake);
